Implement basic lookaround epsilon transitions
**Description**
Implement the `Start` and `End` lookaround epsilon transitions by
adding a block to the epsilon transition functions which will evaluate
the lookaround condition.

If the condition is true, then it will recurse into the epsilon
transition function of the state that was included in the lookaround.
That recursion is guarded by a check on the sparse set so that it won't
continue recursing if no new states have been added.

**Motivation**
Lookaround transitions are a required feature of the `regex` crate.
This commit lays most of the common groundwork using the easiest of the
lookaround transitions.

**Testing Done**
`cargo test`, updated and added new expect tests