Add share regex link and URL parameter loading
This commit adds the ability to share a regex pattern via a URL query
parameter in the web playground.

It introduces a 'Share Regex' button that copies the current
playground URL with the regex as a query parameter to the clipboard.
It also adds logic to read the 'regex' query parameter on page load
and automatically populate and compile the input field.

Disclaimer: This commit was written with assistance from a large
language model (Gemini 2.5 Flash Preview).

Prompts:
 - I'd like the ability to share a URL link that will show a specific
   regex. Can you add: - a new button to the playground that will
   "share" the current regex by generating a URL and copying it to the
   clipboard? - and add functionality so that the presence of a
   query param will prepopulate and precompile a regex?
 - Again, none of your changes have appeared in my local workspace.
   Could you try again?
 - Can you add a link to https://docs.rs/regex/latest/regex/#syntax
   somewhere in the playground so that users can figure out what
   syntax is supported?