Redo public interface of `wahgex` crate
**Description**
 - Upgrade MSRV to 1.87 and edition to 2024
 - Encapsulate `wasmi` engine use in a new module and rework the public
   interface to generate bytecode separate from NFA
 - Add two new cargo features, one for using wasmi as an execution engine
   and the second one to enable compiling regex to bytecode. The
   remaining interface with no features enabled is only able to represent
   bytecode, not execute or compile to it.
 - Add new expect tests for unicode lookaround patterns
 - Update tests to use wasmi execution context

**Motivation**
Overall, this change is intended to be a first draft of the public
interface of this crate. It leaves some room for future execution engines
like `wasmtime`, and it effectively separates compilation from execution.

My idea was that in the future there could be a build-time compilation of
a regex pattern into bytecode, and then at runtime execute the regex
bytecode to do matches, etc. The build-time crate would only need to
enable the `compile` feature of the `wahgex` crate, while the runtime
part would only need a feature that provides the execution engine, like
`wasmi`.

The MSRV and edition change were necessary since some of the functions
in use were only stabilized later on. The added tests are for parity with
ascii lookaround tests.

**Testing Done**
Added new tests, updated consumers to use new public interface.