Optimize branch to transition using br_table
**Description**
Optimize the branch to transition function using the `br_table`
instruction. The `br_table` instruction emulates a `switch` statement
when combined with many `block`s of instructions.

**Motivation**
The previous strategy of `if`-blocks was not efficient, since it resulted
in many individual comparisons. The new `br_table` lets there be a single
comparison that indexes into the list of labels and jumps immediately to
the right transition function.

I tried an alternate form of this that used funcref tables and
`call_indirect`, but that took more space and didn't improve performance
as much.

Part of the reason that this works well for the `branch_to_transition`
function (and not the `branch_to_epsilon` function) is that the majority
of states in an NFA are likely to need a transition, unlike the epsilon
closure. This is especially stark when introducing something like `\w`
into a pattern, which immediately add 300+ states.

I'll include benchmarks on the PR overview.

**Testing Done**
`./scripts/ci.sh`