Add config option to disable WASM module names
**Description**
Add flag to the `Config` struct which allows toggling on and off the
names emitted in the WASM module.

**Motivation**
The name custom section is very helpful for debug scenarios, but takes up
a lot of space relative to the size of the WASM module. On average, half
or more of the total size of the module is just the name section.

Removing it makes the modules faster to load and easier to transport.

Keep names enabled for tests and CLI output because that improves
debugging experience. Also keep the names enabled for the web playground
output to showcase the internals of the WASM module.

**Testing Done**
`./scripts/ci.sh`