Re-order section layouts to allow more compaction
**Description**
Previously the two sparse sets were laid-out in the middle of the other
layout, which could cause a gap between data segments from the
transitions layout and the layouts from patterns and lookarounds.

Changing the order to make them contiguous ensures that all the data
segments will be compacted together and could save a couple of bytes in
most scenarios.

**Testing Done**
`./scripts/ci.sh`